if statements

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -9,21 +9,21 @@
  */
 int main(void)
 {
-	double n;
+	int n;
 
 	srand(time(0));
-	n=rand()-RAND_MAX/2;
+	n = rand() - RAND_MAX / 2;
 	if (n > 0)
 	{
-		printf ("%.1f is positive\n",n);
+		printf ("%d is positive\n", n);
 	}
 	else if (n == 0)
 	{
-		printf ("%.1f is zero\n",n);
+		printf ("%d is zero\n", n);
 	}
 	else if (n < 0)
 	{
-		printf ("%.1f is negative\n",n);
+		printf ("%d is negative\n", n);
 	}
 	return (0);
 }
